requirements.txt
boto3
fastmcp
pydantic
python-dotenv